Bug fixes for flavours and customers
Mike and Tutti Frutti never showed up, because their respective arrays
already contained the values for that customer/flavour respectively.

The Classic and Bob never showed up, because of an incorrect assumption
in how Random.Range works.

Also, of note, the code changes are still compatible with the crueler
variant where the module resets completely on an incorrect sale.